+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Redefined distinct user count for "connected_clients" and added "connected_sockets" for a more accurate count of connected users and sockets. (by @HB9HIL)
+
 ## [0.2.2] - 2026-07-13
 
 ### Changed